@@ -2,7 +2,6 @@
   "name": "example",
   "version": "0.1.0",
   "description": "An example using gatsby-theme-blog-ocean.",
-  "main": "index.js",
   "repository": "https://github.com/HowyoungZhou/gatsby-theme-blog-ocean.git",
   "author": "Howyoung <howyoung3@gmail.com>",
   "license": "MIT",
@@ -13,10 +12,10 @@
     "serve": "gatsby serve"
   },
   "dependencies": {
-    "gatsby": "^4.16.0",
-    "gatsby-source-filesystem": "^4.16.0",
+    "gatsby": "^5.11.0",
+    "gatsby-source-filesystem": "^5.11.0",
     "gatsby-theme-blog-ocean": "0.1.5",
-    "react": "^18.1.0",
-    "react-dom": "^18.1.0"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   }
 }
