@@ -2,7 +2,6 @@
   "name": "gatsby-theme-blog-ocean",
   "version": "0.1.5",
   "description": "A blog theme for Gatsby.",
-  "main": "gatsby-config.js",
   "repository": "https://github.com/HowyoungZhou/gatsby-theme-blog-ocean.git",
   "author": "Howyoung <howyoung3@gmail.com>",
   "license": "MIT",
@@ -22,48 +21,52 @@
     "extract": "yarn run babel --config-file ./babel-extract.config.js -o tmp/chunk.js 'src/**/*.{js,jsx,ts,tsx}' && rm -rf tmp"
   },
   "peerDependencies": {
-    "gatsby": "^4.4.0"
+    "gatsby": "^5.11.0"
   },
   "dependencies": {
-    "@emotion/react": "^11.9.3",
-    "@emotion/styled": "^11.9.3",
-    "@iconify-icons/emojione": "^1.2.2",
-    "@iconify-icons/logos": "^1.2.6",
-    "@iconify/react": "^3.2.2",
-    "@mdx-js/mdx": "^1.6.22",
-    "@mdx-js/react": "^1.6.22",
-    "@mui/icons-material": "^5.8.3",
-    "@mui/lab": "^5.0.0-alpha.85",
-    "@mui/material": "^5.8.3",
-    "gatsby-core-utils": "^3.16.0",
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
+    "@iconify-icons/emojione": "^1.2.6",
+    "@iconify-icons/logos": "^1.2.33",
+    "@iconify/react": "^4.1.1",
+    "@mdx-js/mdx": "^2.3.0",
+    "@mdx-js/react": "^2.3.0",
+    "@mui/icons-material": "^5.14.3",
+    "@mui/lab": "^5.0.0-alpha.140",
+    "@mui/material": "^5.14.5",
+    "gatsby-core-utils": "^4.11.0",
     "gatsby-plugin-feed-mdx": "^1.1.0",
-    "gatsby-plugin-image": "^2.16.1",
-    "gatsby-plugin-manifest": "^4.16.0",
-    "gatsby-plugin-mdx": "^3.16.1",
-    "gatsby-plugin-offline": "^5.16.0",
-    "gatsby-plugin-react-helmet": "^5.16.0",
-    "gatsby-plugin-react-i18next": "^1.2.3",
-    "gatsby-plugin-sharp": "^4.16.1",
-    "gatsby-plugin-typescript": "^4.16.0",
-    "gatsby-remark-copy-linked-files": "^5.16.0",
-    "gatsby-remark-images": "^6.16.0",
-    "gatsby-remark-smartypants": "^5.16.0",
-    "gatsby-source-filesystem": "^4.16.0",
-    "gatsby-theme-material-ui": "^5.2.0",
-    "gatsby-transformer-sharp": "^4.16.0",
-    "i18next": "^21.8.9",
+    "gatsby-plugin-image": "^3.11.0",
+    "gatsby-plugin-manifest": "^5.11.0",
+    "gatsby-plugin-mdx": "^5.11.0",
+    "gatsby-plugin-offline": "^6.11.0",
+    "gatsby-plugin-react-helmet": "^6.11.0",
+    "gatsby-plugin-react-i18next": "^3.0.1",
+    "gatsby-plugin-sharp": "^5.11.0",
+    "gatsby-plugin-typescript": "^5.11.0",
+    "gatsby-remark-copy-linked-files": "^6.11.0",
+    "gatsby-remark-images": "^7.11.0",
+    "gatsby-remark-smartypants": "^6.11.0",
+    "gatsby-source-filesystem": "^5.11.0",
+    "gatsby-theme-material-ui": "^5.3.0",
+    "gatsby-transformer-sharp": "^5.11.0",
+    "i18next": "^23.4.4",
     "path-to-regexp": "^6.2.1",
-    "prism-react-renderer": "^1.3.3",
-    "react": "^18.1.0",
-    "react-dom": "^18.1.0",
+    "prism-react-renderer": "^2.0.6",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
-    "react-i18next": "^11.17.1",
-    "remark-slug": "6.1.0"
+    "react-i18next": "^13.1.2",
+    "rehype-slug": "^5.1.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.17.10",
-    "@babel/plugin-transform-typescript": "^7.18.4",
-    "@types/react-helmet": "^6.1.5",
-    "babel-plugin-i18next-extract": "^0.8.3"
+    "@babel/cli": "^7.22.10",
+    "@babel/plugin-transform-typescript": "^7.22.10",
+    "@types/node": "^20.5.1",
+    "@types/react": "^18.2.20",
+    "@types/react-dom": "^18.2.7",
+    "@types/react-helmet": "^6.1.6",
+    "babel-plugin-i18next-extract": "^0.9.0",
+    "typescript": "^5.1.6"
   }
 }
